Refatorado o código que lida com memória flash Removido o arquivo bin/Arcade_Mania.gba Melhorado arquivo makefile Outras mudanças aleatórias
