+++ include/main.h
@@ -1,0 +1,3 @@
+#ifndef MAIN_H
+#define MAIN_H
+#endif /* MAIN_H */
