+++ include/menu.h
@@ -1,6 +1,8 @@
 #ifndef MENU_H
 #define MENU_H
 
+#include <general.h>
+
 gameIndex_t init_menu(void);
 
 #endif /* MENU_H */
